requirements.txt
apiclient==1.0.4
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
deprecation==2.1.0
Flask==2.0.2
Flask-WTF==1.0.0
google-api-core==2.2.2
google-api-python-client==2.31.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
modzy-sdk==0.6.0
numpy==1.21.4
packaging==21.3
pandas==1.3.4
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
rsa==4.7.2
six==1.16.0
typing_extensions==4.0.0
uritemplate==4.1.1
urllib3==1.26.7
Werkzeug==2.0.2
WTForms==3.0.0
zipp==3.6.0
